Move sub extraction to helper functions
